Better handling of decimals
What seems to be happening is  that when something is computed as 999/1000 for example, the value is still shown as 100% as that is simply how float(x) works. The behavior I am looking for is a round down so more testing will need to be done.

I have substituted with decimal accordingly.